requirements.txt
langchain
langchain-groq
python-dotenv
groq
streamlit